+++ src/BlueBoard.Persistence.Abstractions/Repositories/IUserRepository.cs
@@ -7,9 +7,13 @@ namespace BlueBoard.Persistence.Abstractions.Repositories
     public interface IUserRepository
     {
         Task<bool> IsUserExistsAsync(IDbConnection connection, string email);
+
         Task<UserEntity> CreateUserAsync(IDbConnection connection, string email);
+
         Task<UserEntity> FindByEmailAsync(IDbConnection connection, string email);
-        Task<UserEntity> FindById(IDbConnection connection, int id);
+
+        Task<UserEntity> FindById(IDbConnection connection, long id);
+
         Task<UserEntity> Update(IDbConnection connection, UserEntity entity);
     }
 }
